+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.0+
+
+-Fixing the convertTimeStampToStr(Map<String, dynamic> myMap) to be able to convert TimeStamp values in nested Map
+-You don't need to provide the fields keys you want to convert to TimeStamp inside #convertStrToTimeStamp(Map<String, dynamic> myMap) function .
+